codesmellfix(error component) - moved components out of the parent component
